moved some functionality around
