+++ worldsim/scene.py
@@ -107,3 +107,5 @@ class Scene:
 
         return toprint
     
+
+
